pyproject.toml
[project]
name = "hevy-api"
version = "0.1.0"
description = "Unofficial Python client for Hevy workout data API"
readme = "README.md"
authors = [{name = "Remi Uzel"}]
requires-python = ">=3.10"
dependencies = [
    "cachetools>=6.1.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "argparse>=1.4.0",
    "just>=0.8.163",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.0",
]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "PIE", "SIM", "RET"]
ignore = ["E501"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
